backend/requirements.txt
# --- 核心异步网络框架 ---
aiohttp>=3.9.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# --- 交易所 SDK ---
# GRVT 官方 SDK
grvt-pysdk>=0.2.1

# Lighter.xyz (指定从 GitHub 安装)
lighter-python @ git+https://github.com/elliottech/lighter-python.git

# 通用依赖
ccxt>=4.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
web3>=6.0.0